Can set Concurrent::Executor for returned Future
Actually, this change permits overriding any of the constructor args on
the returned Concurrent::Future when running an RPC, including
providing an :executor.

When constructing the Elementary::Connection, can can provide a hash of
:future_options that are passed down into the Concurrent::Future
constructor.  This makes it easy to get these operations off of the
global thread pools if needed.