Create Http client object in constructor
This fixes issue #43. If the client object is not created in the
contructor (which is called immediately when the Elementary::Rpc
connection is created), then if multiple threads use that connection,
they can run into races. Multiple threads could modify @options in
parallel, or one thread could use @client (which freezes the adapater
list) while another thread was still initializing it).

This also fixes a small bug in the connection spec, which was passing
invalid options in, but were never checked previously. I'm bumping minor
version and not patch for this reason -- there is a slight change in
behavior in this regard.